+++ CHANGELOG.md
@@ -9,6 +9,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 * Added `__json__` to sensorchan specs
+* Added `resize` method
+
+### Fixed
+* A Dequantize node between a warp and and overview now has its size modified correctly.
+
+### Changed
+* `write_network_text` now defaults to rich='auto'
 
 
 ## [Version 0.2.0] - Released 2022-09-27